fix:指摘対応
