+++ src/features/auth/components/AuthButton.tsx
@@ -1,12 +1,12 @@
 import { Button } from "@/components/ui/button";
 import { Loader2 } from "lucide-react";
 
-type LoginButtonProps = {
+type AuthButtonProps = {
 	label: string;
 	isPending: boolean;
 };
 
-const LoginButton = ({ label, isPending = false }: LoginButtonProps) => {
+const AuthButton = ({ label, isPending = false }: AuthButtonProps) => {
 	return (
 		<Button type="submit" className="w-full" disabled={isPending}>
 			{isPending ? <Loader2 className="animate-spin" /> : label}
@@ -14,4 +14,4 @@ const LoginButton = ({ label, isPending = false }: LoginButtonProps) => {
 	);
 };
 
-export default LoginButton;
+export default AuthButton;
